Delete Alert

+++ NEW/resident.php
@@ -324,7 +324,7 @@ $result=mysqli_query($mysqli, "SELECT * FROM residents");
                                 echo '<tr>';
 
                                   echo '<td><button type="button" class="btn btn-primary editbtn">Edit</td>'; 
-                                 echo '<td><a class="btn btn-danger" href="delete.php?id='. $res['id'] . '" role="button">Delete</a></td>';
+                                echo '<td><a onclick="confirmationDelete(this)" class="btn btn-danger" href="delete.php?id='. $res['id'] . '" role="button ">Delete</a></td>';
                                   echo '<td>'. $res['id'] .'</td>' ;
                                   echo '<td>'.'<img src="data:images/jpeg;base64,'.base64_encode(stripslashes($res['images'])).'"/>' .'</td>';
                                   echo '<td>'. $res['lastName'] .'</td>' ;
@@ -1105,6 +1105,15 @@ $result=mysqli_query($mysqli, "SELECT * FROM residents");
         
         
     </script>
+      <!----CONFIRMATION TO DELETE--->
+      <script>
+          function confirmationDelete(anchor)
+{
+   var conf = confirm('Are you sure want to delete this record?');
+   if(conf)
+      window.location=anchor.attr("href");
+}
+      </script>
 
 </body>
